+++ packages/table/src/filters/UserFilter/UserFilter.tsx
@@ -1,10 +1,6 @@
 import * as React from 'react';
 
-import {
-  SelectControl,
-  type TSelectOption,
-  Avatar,
-} from '@devoinc/genesys-ui';
+import { SelectControl, type TSelectOption, Avatar } from '@devoinc/genesys-ui';
 
 import type { TFilterContext, TFilter } from '../../declarations';
 import type { TUserFilterValue } from './declarations';
@@ -22,6 +18,7 @@ export const UserFilter: React.FC<TFilter> = ({ colDef, onChange }) => {
         colorScheme={user?.colorScheme || 'info'}
         size="xxxs"
         name={user?.name}
+        initials={user?.initials}
       />
     );
   };
@@ -34,8 +31,9 @@ export const UserFilter: React.FC<TFilter> = ({ colDef, onChange }) => {
       } else {
         return {
           value: keys[index],
-          label: user.name,
+          label: user.isCurrentUser ? 'Current user' : user.name,
           icon: <AvatarCmp user={user} />,
+          tooltip: user.email,
         } as TSelectOption;
       }
     },