+++ packages/table/src/renderers/BulkRenderer/BulkRendererById.tsx
@@ -2,13 +2,17 @@ import * as React from 'react';
 
 import { CheckboxControl } from '@devoinc/genesys-ui';
 
-import type { TCellRenderer } from '../../declarations';
+import type { TColDef } from '../../declarations';
 import type { TBulkContext } from '../../facade';
 
-export const BulkRendererById: React.FC<TCellRenderer> = ({
+export const BulkRendererById: React.FC= ({
   value,
   rowIndex,
   colDef,
+} : {
+  value: boolean,
+  rowIndex: number,
+  colDef: TColDef
 }) => {
   const context = colDef?.context as TBulkContext;
   return (