+++ packages/table/src/renderers/UserRenderer/declarations.ts
@@ -10,6 +10,8 @@ export type TContextUser = {
       job?: string;
       role?: string;
       subtitle?: string;
+      initials?: string;
+      isCurrentUser?: boolean;
     };
   };
 };